feat: initialize save book on db

+++ src/main/java/com/biblioteca/gp5/integration/gutendex/client/GutendexClient.java
@@ -33,7 +33,7 @@ public class GutendexClient {
 						.body(GutendexSearchResponseDTO.class);
 	}
 	
-	public GutendexBookResponseDTO searchBookById(Long id) {
+	public GutendexBookResponseDTO searchBookById(Integer id) {
 		
 		return restClient.get()
 						.uri("/books/{id}", id)
